+++ CHANGELOG.md
@@ -16,6 +16,7 @@ As this project is a user-facing application, the places in the semantic version
 ### Added
 
 - Docker support
+- Show year tooltip when dragging year filter slider handles
 
 ## [2.2.7] (2023-11-29)